Moved types and constants to separate files per feedback.
