+++ server/src/constants.ts
@@ -3,3 +3,10 @@ export const BUNDLE_TYPE = {
   AMENDMENT: "AMENDMENT",
   EXTENSION: "EXTENSION",
 } as const;
+
+export const SIGNATURE_LEVEL = ["OA", "OCD", "OGD"] as const;
+
+export const CMCS_DIVISION = [
+  "Division of System Reform Demonstrations",
+  "Division of Eligibility and Coverage Demonstrations",
+] as const;
