+++ server/src/types.ts
@@ -1,4 +1,4 @@
-import { BUNDLE_TYPE } from "./constants.js";
+import { BUNDLE_TYPE, CMCS_DIVISION, SIGNATURE_LEVEL } from "./constants.js";
 
 // Export types for use in the client code
 export type {
@@ -66,3 +66,5 @@ export type {
 export type { Document } from "./model/document/documentSchema.js";
 
 export type BundleType = (typeof BUNDLE_TYPE)[keyof typeof BUNDLE_TYPE];
+export type CmcsDivision = (typeof CMCS_DIVISION)[number];
+export type SignatureLevel = (typeof SIGNATURE_LEVEL)[number];
